# api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Internal HTTP API",
    "version": "1.0.0-alpha",
    "summary": "Public reader endpoints, editorial writes behind a bearer JWT, and the operational probes served by the Go binary.",
...
  },
  "servers": [
    {
      "url": "http://api:8080",
      "description": "The internal service address inside compose and Kubernetes; the Astro server reaches it as API_BASE_URL."
    },
    {
      "url": "http://localhost:8080",
      "description": "A locally run binary (HTTP_ADDR defaults to :8080)."
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/cashback/export": {
      "get": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "exportCashbackHistory",
        "summary": "Export the caller's own cashback history",
        "description": "The member's complete cashback history as one document, in JSON or CSV (FR-003).\n\nFR-003 pairs leaving with taking your record, and for a reason: a member deciding whether to go should not have to choose between the two.\n\nThis is NOT a page. GET /wallet/entries pages because a member reads their history a screen at a time; an export is read by a spreadsheet, an accountant or a subject-access request, and a partial one is worse than none - a document that looks complete and is not. So there is no state, limit or cursor, and sending one is a 400 rather than a narrower file.\n\nThe CSV carries each money figure as TWO columns, minor units and currency, because C-6 holds in a spreadsheet exactly as on the wire: one column of bare integers is a column somebody sums across currencies. Text a spreadsheet would read as a formula is prefixed with an apostrophe - merchant names come from an affiliate network's catalogue and reasons are typed by operators, and neither is text this service wrote.\n\nBoth renderings are sent as an attachment, newest entry first.\n\nThe member is the bearer of the token. There is no path segment and no query parameter naming an account, so exporting somebody else's history is not a request this endpoint can express.\n\nThe whole path is unmounted when cashback is off, and then answers 404.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "json (the default) or csv.",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ],
              "default": "json"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "description": "BCP-47 primary language subtag to name merchants in. A name shown in another language is labelled as a fallback rather than passed off (US5 scenario 2). Omitted, every name comes back labelled.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The member's complete history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashbackExport"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string",
                  "description": "A header row followed by one row per entry, newest first: entry_id, transacted_at, merchant_name, merchant_name_language, merchant_name_is_fallback, sale_amount_minor, sale_currency, cashback_amount_minor, cashback_currency, state, hold_rule, reversal_of_id, reason. An absent value is an empty cell."
                }
              }
            }
          },
          "400": {
            "description": "An unknown query parameter, a format that is not json or csv, or a state, limit or cursor - which this endpoint refuses by name, because an export is always the complete history.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "413": {
            "description": "This history holds more entries than one document can carry. Refused rather than shortened: a CSV cut short is indistinguishable from a complete shorter one, so a member handed one would have a file they believe is their record and is not.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A JWT from the auth provider, verified against the configured JWKS. The subject must map to an account. Endpoints under /api/v1/editorial/ additionally require that account to hold the editor role, and the database enforces that rule again on write; endpoints under /api/v1/account/ require only that somebody is signed in, because they act on the caller's own row."
      }
    },
...
    "responses": {
      "Unauthenticated": {
        "description": "No bearer token, or one that is invalid, expired or belongs to no account.",
        "headers": {
          "WWW-Authenticate": {
            "description": "`Bearer realm=\"editorial\"`, with `error=\"invalid_token\"` appended when a token was supplied but rejected.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "missingToken": {
                "value": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "a bearer token is required"
                }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The request failed for a reason the caller cannot act on. Internals stay in the log: the body carries no database or stack detail, and `detail` may be absent entirely.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "opaque": {
                "value": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "title": "Problem",
        "description": "RFC 9457 problem details, as every handler in this API writes them.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "description": "Always `about:blank`, which per RFC 9457 makes `title` the plain rendering of the status code. No problem-type URIs are minted for this API.",
            "const": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "The HTTP status text, e.g. `Bad Request`.",
            "examples": [
              "Bad Request"
            ]
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code, repeated in the body.",
            "minimum": 400,
            "maximum": 599,
            "examples": [
              400
            ]
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation of this occurrence. Omitted when the server has nothing safe to say (notably on 500)."
          }
        }
      },
...
      "Money": {
        "type": "object",
        "description": "An amount of money: whole minor units and the ISO-4217 code they are denominated in. Never a decimal and never a float - the frontend formats these two fields for a locale, and nothing between here and there has to agree on how many decimal places a currency has (C-6).",
        "required": [
          "minor",
          "currency"
        ],
        "additionalProperties": false,
        "properties": {
          "minor": {
            "type": "integer",
            "format": "int64",
            "description": "The amount in minor units, e.g. 4999 for 49.99 EUR."
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "example": "EUR"
          }
        }
      },
...
      "WalletEntry": {
        "type": "object",
        "description": "One line of a member's history: the purchase, what it earned, and where that stands.",
        "required": [
          "entry_id",
          "merchant_name",
          "merchant_name_language",
          "merchant_name_is_fallback",
          "transacted_at",
          "sale_amount",
          "cashback_amount",
          "state",
          "expected_confirmation_at",
          "hold_rule",
          "reversal_of_id",
          "reason"
        ],
        "additionalProperties": false,
        "properties": {
          "entry_id": {
            "type": "string",
            "format": "uuid"
          },
          "merchant_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The retailer, or null for an entry an operator attributed by hand: there was no click, so there is no route to a retailer to follow (FR-034)."
          },
          "merchant_name_language": {
            "type": [
              "string",
              "null"
            ],
            "description": "The BCP-47 primary subtag merchant_name is written in, or null when there is no name."
          },
          "merchant_name_is_fallback": {
            "type": "boolean",
            "description": "True when merchant_name is in the retailer's own language rather than the one asked for. US5 scenario 2 requires that to be shown as a fallback rather than passed off as the member's language."
          },
          "transacted_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the purchase happened, as the network reported it - the instant a member recognises. Not when the entry was written."
          },
          "sale_amount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              }
            ],
            "description": "What the member spent."
          },
          "cashback_amount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              }
            ],
            "description": "What they earned on it."
          },
          "state": {
            "type": "string",
            "enum": [
              "held",
              "pending",
              "confirmed",
              "reserved",
              "paid",
              "reversed"
            ],
            "description": "Where this entry sits now (FR-042)."
          },
          "expected_confirmation_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "When this entry is expected to confirm. ALWAYS null today: nothing in the schema records a confirmation window, so there is no honest value to compute, and an invented one is a date a member would plan around. It becomes a real answer when a window is recorded per route.",
            "format": "date-time"
          },
          "hold_rule": {
            "type": [
              "string",
              "null"
            ],
            "description": "The rule holding this entry back, and null unless the state is held (US7). The wallet totals do not show held money at all, so a member looking at an entry they cannot count is owed the reason."
          },
          "reversal_of_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The entry this one undoes, and null on an entry that undoes nothing (SC-010).",
            "format": "uuid"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why this entry exists, recorded with the transition that opened it. On a reversal, why the money went back."
          }
        }
      },
...
      "CashbackExport": {
        "type": "object",
        "description": "A member's complete cashback history as one document (FR-003). Not a page: an export is the whole record or it is not an export, so this endpoint takes no state, limit or cursor.",
        "required": [
          "exported_at",
          "entry_count",
          "entries"
        ],
        "additionalProperties": false,
        "properties": {
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this document was taken. The export walks the history backwards, so an entry written while it was being assembled is newer than the pages already read and is not in the file; this is what lets a reader tell what the document covers."
          },
          "entry_count": {
            "type": "integer",
            "minimum": 0,
            "description": "How many entries follow. Written out rather than left to be counted, so a truncated file can be DETECTED: the count and the array would disagree."
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WalletEntry"
            },
            "description": "The whole history, newest first - the same order and the same item shape as GET /wallet/entries, so a client parses one thing for both."
          }
        }
      },
...
    }
  }
}